requirements.txt
asgiref==3.7.2
Django==4.2.3
django-debug-toolbar==4.1.0
django-js-asset==2.1.0
django-mptt==0.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
